Export dist file instead of source.
We shouldn't expect other projects to build angular-tootips since the
build process is specific to the package like using annotation and so
on.
If this is required by a project that doesn't use annotation plugin, it
will cause Angular to throw `strict di` error message. By exporting the
dist file, it will be ready to use right away.